Prepare inspector for non-`hdf5_type*` iterators.
Improving the string support requires that (de-)serialization can
happen from an iterator, as opposed to the current `hdf5_type*`.

This commit prepares the containers which largely just move the pointer
forward by the correct number of elements and hand it over to the next
serialization routine of its elements, so that they can accept something
pointer like, i.e. an iterator.